Add CircleCI support for user ruby
On MacOS 10.15, we need to use a newer version of ruby than the system
version.  To accomplish this, we install a user instance of ruby and set
up the environment in .bash_profile.  CircleCI will launch following
scripts with a new instance of bash, so these variables will be picked
up at that time.